fix(runners): the five the code round accepted — and the ledger lands where no test has to move
The gate read the diff (nine reviewers) and returned `revise`. Five findings were worth
taking.

Codex caught that the ledger's own test had gained a `using`, which contradicts this
story's whole claim — every existing test unedited is the proof a move was a move. The fix
is not to reword the claim: UsageLedger now lives in CoaiMcp.Runners.Reviewers, beside the
outcomes it records, which is a namespace that test already imports. It is byte-identical
to main again. (Runners.Usage, the first attempt, shadowed the `Usage` TYPE across the
whole library; Runners.Accounting compiled and would have cost exactly the edit above.)

Gemini found the note that helps nobody: reading stderr alone left "--version exited 1: "
for every CLI that diagnoses itself on stdout — an exit code, a colon and nothing. The note
now carries whichever stream the vendor actually used, with a test that pins it.

The probe also caught only Win32Exception, so a path that names a directory, or a handle
that will not open, faulted the whole catalog answer rather than describing one vendor.
IOException and InvalidOperationException join it: a health probe that throws is a worse
answer than any note.

And two small ones: the runtime is resolved once per probe rather than twice — `For`
constructs an adapter, so the second call allocated one per vendor per catalog call for
nothing — and CliFound's contract is written down, because it means the executable STARTED,
not that it works. A CLI that started and then went silent is `true` with a note, and the
two failures have different cures.

Seventeen were rejected with reasons. Nine of them were one argument from the local model —
make every new type `internal` — which is backwards for the only change this story makes:
these types exist so a SECOND ASSEMBLY can ask the same questions instead of keeping a
fourth copy, and internal is per-assembly. Two more filed the same "the model is lost"
finding twice, when the model is a parameter of RunAsync, is checked in LocalHealth and has
a passing test on that arm — its own text breaks off with "Wait, re-". One asked for a UI
spinner in a change that moves code between two C# projects, and the panel's missing
probing state is already its own open plan.

773 tests, 772 green, 1 skipped.
