+++ src_mcp/src/Server/PanelService.cs
@@ -7,7 +7,6 @@ using CoaiMcp.Core.Rounds;
 using CoaiMcp.Runners.Context;
 using CoaiMcp.Runners.Processes;
 using CoaiMcp.Runners.Reviewers;
-using CoaiMcp.Runners.Accounting;
 using CoaiMcp.Runners.Worktrees;
 
 namespace CoaiMcp.Server;
