+++ src_mcp/src/Server/ServerJsonContext.cs
@@ -1,7 +1,6 @@
 using System.Text.Json.Serialization;
 using CoaiMcp.Core.Findings;
 
-using CoaiMcp.Runners.Accounting;
 
 namespace CoaiMcp.Server;
